Include the awesome-pages plugin
Fix plugin configuration and installation by:
- Explicitly installing used plugins in the `publish-site.yml` workflow
- Moving the `plugins:` YML block to the root of the `mkdocs.yml` file

Install the **awesome-pages** plugin, and add a sample `.pages` file in the root of the `docs` folder, which allows configuration of page nav order.

If approved, a future PR could then rename all the folders to remove the numbering, and add in `.pages` files where necessary to preserve desired ordering.